updated logic for accessing /chat when logged in and no chatrooms
